vmh264: fix video jitter bug [1/1]
PD#SWPL-24276

Problem:
Interlaced stream jitter.

Solution:
This type of stream consists of a pack of two fields
and a pack of single fields.
When the second field is the same as the pts of the first field,
the display module will drop the second field, causing jitter.
The pts of the second field needs to be set to 0,
and the display module reconstructs the pts
through duration / 2 to output the second field.

Verify:
u212

Change-Id: I53b0c4104c5e21a98423e49bb27831b1e9b20e11
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>